# Up to 30x speedup for 2D plotting