--- a/python/src/scipp/plot/plot_2d.py
+++ b/python/src/scipp/plot/plot_2d.py
@@ -16,24 +16,6 @@ import ipywidgets as widgets
 import matplotlib.pyplot as plt
 import warnings
 
-import os
-import time
-
-
-def timeit(method):
-    def timed(*args, **kw):
-        ts = time.time()
-        result = method(*args, **kw)
-        te = time.time()
-        if 'log_time' in kw:
-            name = kw.get('log_name', method.__name__.upper())
-            kw['log_time'][name] = int((te - ts) * 1000)
-        else:
-            os.write(1, f'{method.__name__}  {(te - ts) * 1000} ms\n'.encode())
-        return result
-
-    return timed
-
 
 def plot_2d(scipp_obj_dict=None,
             axes=None,
@@ -213,7 +195,6 @@ class Slicer2d(Slicer):
         self.update_axes()
         return
 
-    @timeit
     def update_axes(self):
         # Go through the buttons and select the right coordinates for the axes
         for dim, button in self.buttons.items():
@@ -304,7 +285,6 @@ class Slicer2d(Slicer):
 
         return
 
-    @timeit
     def compute_bin_widths(self, xy, dim):
         """
         Pixel widths used for scaling before rebin step
@@ -313,7 +293,6 @@ class Slicer2d(Slicer):
                             self.xyedges[xy][dim, :-1])
         self.xywidth[xy].unit = sc.units.one
 
-    @timeit
     def slice_coords(self):
         """
         Recursively slice the coords along the dimensions of active sliders.
@@ -338,7 +317,6 @@ class Slicer2d(Slicer):
             # Pixel widths used for scaling before rebin step
             self.compute_bin_widths(xy, param["dim"])
 
-    @timeit
     def slice_data(self):
         """
         Recursively slice the data along the dimensions of active sliders.
@@ -378,7 +356,6 @@ class Slicer2d(Slicer):
         self.vslice *= self.xywidth["x"]
         self.vslice *= self.xywidth["y"]
 
-    @timeit
     def update_slice(self, change=None):
         """
         Slice data according to new slider value and update the image.
@@ -408,7 +385,6 @@ class Slicer2d(Slicer):
         if self.xlim_updated:
             self.update_bins_from_axes_limits()
 
-    @timeit
     def update_bins_from_axes_limits(self):
         """
         Update the axis limits and resample the image according to new viewport
@@ -457,7 +433,6 @@ class Slicer2d(Slicer):
         last = min(bins - 1, last)
         return dim, slice(first, last + 1)
 
-    @timeit
     def resample_image(self):
         dim = self.xyrebin['x'].dims[0]
         slicex = self.select_bins(self.xyedges['x'], dim,
@@ -499,7 +474,6 @@ class Slicer2d(Slicer):
         arr /= self.xyrebin['y'].values[1] - self.xyrebin['y'].values[0]
         return arr
 
-    @timeit
     def update_image(self, extent=None):
         dslice = self.resample_image()
         if self.params["masks"][self.name]["show"]: