--- a/python/src/scipp/plot/plot_2d.py
+++ b/python/src/scipp/plot/plot_2d.py
@@ -364,7 +364,7 @@ class Slicer2d(Slicer):
             self.vslice = sc.histogram(self.vslice)
             self.autoscale_cbar = True
         else:
-            self.vslice = self.vslice.copy()
+            self.vslice = self.vslice.astype(sc.dtype.float32)
             self.autoscale_cbar = False
         self.vslice.variances = None
         self.vslice.unit = sc.units.counts
@@ -445,12 +445,14 @@ class Slicer2d(Slicer):
         return
 
     def select_bins(self, coord, dim, start, end):
-        if len(coord.dims) != 1:
+        if len(coord.dims) != 1: # TODO find combined min/max
             return dim, slice(0, -1)
         bins = coord.shape[0]
         # scipp treats bins as closed on left and open on right: [left, right)
         first = sc.sum(coord <= start, dim).value - 1
         last = bins - sc.sum(coord > end, dim).value
+        if first >= last: # TODO better handling for decreasing
+            return dim, slice(0, -1)
         first = max(0, first)
         last = min(bins - 1, last)
         return dim, slice(first, last + 1)