--- a/python/src/scipp/plot/plot_2d.py
+++ b/python/src/scipp/plot/plot_2d.py
@@ -427,7 +427,7 @@ class Slicer2d(Slicer):
     def select_bins(self, coord, dim, start, end):
         bins = coord.shape[-1]
         if len(coord.dims) != 1:  # TODO find combined min/max
-            return dim, slice(0, bins-1)
+            return dim, slice(0, bins - 1)
         # scipp treats bins as closed on left and open on right: [left, right)
         first = sc.sum(coord <= start, dim).value - 1
         last = bins - sc.sum(coord > end, dim).value