### Example:
```python
import scipp as sc
import numpy as np
# This is roughly LoKI with full pixel count, but bins might also be 1000 in a worse case
data = sc.DataArray(data=sc.Variable(dims=['x','y'], values=np.random.rand(1000000,100)))
data.masks['m'] = data.data['x',0] < data.data['x',1]
```
```python
%%time
sc.plot.plot(data)
```

Compiled with TBB.

### Before

- `35 s` for the plot to show up.
- `1 s` to `3 s` for zoom into fewer pixels
- `1 s` to `3 s` for pan in zoomed-in state

### After

- `1 s` for the plot to show up (there must be extra overhead outside the 2d plotting code, the update takes only `200 ms`).
- `50 ms` for zoom into fewer pixels
- `50 ms` for pan in zoomed-in state

### How?

- Copy slice once, early, instead of once for every update.
- Optimize algorithm in `rebin_non_inner`.
- Multi-threading in `rebin_non_inner`.
- Slice before rebinning to avoid rebinning areas that are not required.