@nvaytet 2020-08-14
Something doesn't seem quite right.
Playing around with this on my large image example, i do see the speedup, but the data starts to look wrong when i zoom in.
The code example is:
```Py
N = 6000
M = 3000
xx = np.arange(N, dtype=np.float64)
yy = np.arange(M, dtype=np.float64)
x, y = np.meshgrid(xx, yy)
b = N/200.0
c = M/20.0
r = np.sqrt(((x-c)/b)**2 + (y/b)**2)
a = np.sin(r)
d1 = sc.Dataset()
d1.coords['x'] = sc.Variable(['x'], values=-xx, unit=sc.units.m)
d1.coords['y'] = sc.Variable(['y'], values=yy, unit=sc.units.m)
d1['Signal'] = sc.Variable(['y', 'x'], values=a + np.random.random([M, N]), unit=sc.units.counts)
p1 = plot(d1)
```
The initial view looks ok with both this branch and `master`:
![Screenshot at 2020-08-14 12-54-33](https://user-images.githubusercontent.com/39047984/90242428-52119500-de2d-11ea-92d3-23f009da4d64.png)
but when i zoom in, the data values seem to change with this branch (there is much more yellow)
![Screenshot at 2020-08-14 12-45-47](https://user-images.githubusercontent.com/39047984/90242477-69e91900-de2d-11ea-9e6a-10beec8d1465.png)
Additionally, I see some horizontal striping that does not appear on the `master` plot:
![Screenshot at 2020-08-14 12-52-46](https://user-images.githubusercontent.com/39047984/90242526-81280680-de2d-11ea-8b83-51ac37b75b8c.png)