@nvaytet 2020-08-14
Yes, this now works.

Plotting with decreasing coordinates seems to miss the last bin?
![Screenshot at 2020-08-14 15-26-58](https://user-images.githubusercontent.com/39047984/90254602-52b52600-de43-11ea-952b-d89debd45ec6.png)

Master:
![Screenshot at 2020-08-14 15-31-09](https://user-images.githubusercontent.com/39047984/90254622-59439d80-de43-11ea-8d35-c71420274263.png)