@nvaytet 2020-08-14
Also seems something is wrong with the Mantid config for the linux builder that builds the docs. Something to do with Poco?

https://dev.azure.com/scipp/scipp/_build/results?buildId=1210&view=logs&j=f2a3abc3-f415-54ca-77f6-7b36a30e229e&t=80d1a925-4103-5c17-e9be-9fe209d0b68c&l=243