@SimonHeybrock 2020-08-17
> Plotting with decreasing coordinates seems to miss the last bin?

Fixed (same for identical problem with 2d coord).