@nvaytet 2020-08-18
We could also get a good speedup by changing the default resolution to something lower, e.g. 400x267 (half of the current).
You can see this easily with `plot(..., resolution=[400, 267])`.
The tradeoff is that you would then start to see rebin artifacts between pixels. But if you really need performance for a specific case, this can be a quick temporary workaround.
This should probably be documented.
I should go through all the different plotting arguments a list them and what they do in the documentation.