@SimonHeybrock 2020-08-18
Are you sure this would yield a speedup? If I remember my final benchmarks the majority of time was *not* spent on work related to the output size, but rather the input size.